Create README.md
Here are 3 simple projects using Python. This demonstrates my skills and interest in the following areas: The first project is a password generator, the second is a QR code generator, and the third is a password checker. To open it, you need to have a Python editor, import some modules, and then simply run the code.