Create password.txt
This straightforward program checks whether your password is included in the list of most common passwords. To use the program, you should download the list of most common passwords from the internet and then check if your password is on that list. You also need to replace "password.txt" with your file name in the line "with open('password.txt', 'r') as file:".